+++ examples/cvode/fcmix_serial/fcvRoberts_dnsL.f
@@ -36,7 +36,7 @@ C
       INTEGER METH, ITMETH, ITOL, ITASK, JOUT, NOUT, IERROOT
       INTEGER INFO(2)
 C The following declaration specification should match C type long int.
-      INTEGER*4 NEQ, IOUT(25), IPAR
+      INTEGER*8 NEQ, IOUT(25), IPAR
 C The following declaration specification should match C type int.
       INTEGER*4 NEQL
       DOUBLE PRECISION RTOL, T, T0, TOUT
@@ -173,7 +173,7 @@ C Fortran routine for right-hand side function.
       IMPLICIT NONE
 C
 C The following declaration specification should match C type long int.
-      INTEGER*4 IPAR(*)
+      INTEGER*8 IPAR(*)
       INTEGER IER
       DOUBLE PRECISION T, Y(*), YDOT(*), RPAR(*)
 C
@@ -194,7 +194,7 @@ C Fortran routine for root finding
 C
       DOUBLE PRECISION T, Y(*), G(*), RPAR(*)
 C The following declaration specification should match C type long int.
-      INTEGER*4 IPAR(*)
+      INTEGER*8 IPAR(*)
       INTEGER IER
 C
       G(1) = Y(1) - 1.0D-4
@@ -213,7 +213,7 @@ C Fortran routine for dense user-supplied Jacobian.
       IMPLICIT NONE
 C
 C The following declaration specification should match C type long int.
-      INTEGER*4 N, IPAR(*)
+      INTEGER*8 N, IPAR(*)
       INTEGER IER
       DOUBLE PRECISION T, Y(*), FY(*), JAC(N,*), H, RPAR(*)
       DOUBLE PRECISION V1(*), V2(*), V3(*)
